+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 These are well.. Changelogs for Animating.Rocks-PLUS that go back to the very start.
 
+### Version 1.1
+
+#### Released on 2/3/26
+
+Added:
+
+- Configurable keybinds
+- Settings page (mayby used in the future for more features)
+
 ### Version 1.0
 
 #### Released on 15/2/26